+++ zephyrus-macros/src/command/details.rs
